+++ CHANGES.rst
@@ -8,6 +8,7 @@ Unreleased Changes
 * `Issue #421 <https://github.com/jantman/awslimitchecker/issues/421>`__
 
   * Stop referencing deprecated ``botocore.vendored.requests.exceptions.ConnectTimeout`` in favor of new, and higher-level, ``botocore.exceptions.ConnectionError``
+  * In :py:meth:`awslimitchecker.utils._get_latest_version`, replace use of ``botocore.vendored.requests`` with ``urllib3``.
 
 7.0.0 (2019-08-13)
 ------------------